Update plugins.adoc
Very minor editing to improve clarity (some may be subjective, like making file names bold, but it matches the other KiCad documentation I've seen). Gave the two tutorial sections (basic but descriptive) names instead of nondescript numbers. Fixed a few typos. Removed "shall" which was suddenly used and out-of-place with the rest of the plebian writing style. Note that I did not check any of the code.

Issues that need reviewing:
1. On line 539 I assume there's only one dialog for selection files, but maybe that's wrong?
2. There are references which don't work. Are these in another format and need to be changed to asciidoc? Does anyone know?
3. Different formatting for file names, classes, functions, etc. Maybe this is good, or maybe there's already a standard I haven't found?