Update packages to install for Debian/Ubuntu
The old list of needed packages isn't reflecting the current state for
Debian Stable (Stretch) / Testing and Ubuntu 18.04 as well.
For Spanish documentation there is also texlive-lang-spanish needed, if
texlive-lang-all is used. texlive-lang-dutch isn't existing in current
Debian based distributions so remove this package from the list and use
the successor package texlive-lang-european.