Remove package name mangling for SINGLE_LANGUAGE option
SINGLE_LANGUAGE is deprecated in favor of the LANGUAGES option. When using
multiple languages, adding these to the package file name would end up
in something like "kicad-doc-de_en_it-5.0.0". As it is questionable if
the combination of specifying LANGUAGES and building a tar.gz package
is ever used just drop the special handling for the package name.